cp-78: * move MOBILE_DIMENSION to media-query.constant.ts file
