+++ frontend/src/libs/constants/constants.ts
@@ -1,1 +1,2 @@
+export { MOBILE_DIMENSION } from './media-query.constant.js';
 export { EMPTY_ARRAY_LENGTH, FIRST_ARRAY_INDEX } from 'shared/build/index.js';
