+++ frontend/src/libs/constants/index.constant.ts
@@ -1,6 +1,4 @@
 const FIRST_INDEX = 0;
 const LAST_INDEX = -1;
 
-const MOBILE_DIMENSION = '(max-width: 767px)';
-
-export { FIRST_INDEX, LAST_INDEX, MOBILE_DIMENSION };
+export { FIRST_INDEX, LAST_INDEX };
