+++ frontend/src/libs/constants/media-query.constant.ts
@@ -1,0 +1,3 @@
+const MOBILE_DIMENSION = '(max-width: 767px)';
+
+export { MOBILE_DIMENSION };
